+++ internal/pkg/runtime/engine/singularity/container_linux.go
@@ -11,7 +11,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"io"
 	"os"
 	osuser "os/user"
 	"path/filepath"
@@ -1561,7 +1560,7 @@ func (c *container) addHomeStagingDir(system *mount.System, source string, dest 
 
 	bindSource := !c.engine.EngineConfig.GetContain() || c.engine.EngineConfig.GetCustomHome()
 
-	// use the session home directory is the user home directory doesn't exist (issue #4208)
+	// use the session home directory if the user home directory doesn't exist (issue #4208)
 	if _, err := os.Stat(source); os.IsNotExist(err) {
 		bindSource = false
 	}
@@ -1575,7 +1574,25 @@ func (c *container) addHomeStagingDir(system *mount.System, source string, dest 
 		system.Points.AddRemount(mount.HomeTag, homeStage, flags)
 		c.session.OverrideDir(dest, source)
 	} else {
-		sylog.Debugf("Using session directory for home directory")
+		// c.engine.EngineConfig.GetContain() is true here unless
+		// the home directory didn't exist.  We could let it always
+		// use the workdir if workdir was defined, but that might be
+		// surprising so check again.
+		workdir := c.engine.EngineConfig.GetWorkdir()
+		if workdir != "" && c.engine.EngineConfig.GetContain() {
+			sylog.Debugf("Using work directory for home directory")
+			workdir, err := filepath.Abs(filepath.Clean(workdir))
+			if err != nil {
+				return "", fmt.Errorf("can't determine absolute path of workdir %s: %s", workdir, err)
+			}
+
+			homeStage = filepath.Join(workdir, "home")
+			if err := fs.Mkdir(homeStage, 0o700); err != nil && !os.IsExist(err) {
+				return "", fmt.Errorf("failed to create %s: %s", homeStage, err)
+			}
+		} else {
+			sylog.Debugf("Using session directory for home directory")
+		}
 		c.session.OverrideDir(dest, homeStage)
 	}
 
@@ -2245,20 +2262,65 @@ func (c *container) addResolvConfMount(system *mount.System) error {
 	resolvConf := "/etc/resolv.conf"
 
 	if c.engine.EngineConfig.File.ConfigResolvConf {
+		skipBinds := c.engine.EngineConfig.GetSkipBinds()
+		skipAllBinds := slices.Contains(skipBinds, "*")
+		if slices.Contains(skipBinds, resolvConf) || skipAllBinds {
+			sylog.Verbosef("Skipping bind of the host's %s", resolvConf)
+			return nil
+		}
+
 		var err error
 		var content []byte
+		dest := resolvConf
+		flags := uintptr(syscall.MS_BIND)
 
 		dns := c.engine.EngineConfig.GetDNS()
 
 		if dns == "" {
-			r, err := os.Open(resolvConf)
+			var fi os.FileInfo
+			fi, err = os.Lstat(resolvConf)
 			if err != nil {
 				return err
 			}
-			content, err = io.ReadAll(r)
-			r.Close()
-			if err != nil {
-				return err
+			done := false
+			if fi.Mode()&os.ModeSymlink != 0 {
+				target, err := os.Readlink(resolvConf)
+				if err != nil {
+					return err
+				}
+				// If the symlink points to a file under /run, for example one made by
+				// systemd-resolved, then copy the symlink and bind in the parent
+				// directory of the file it points to as well.
+				if target[0:5] == "/run/" {
+					dst := filepath.Join(c.session.Layer.Dir(), resolvConf)
+					if err = c.session.AddSymlink(dst, target); err != nil {
+						return fmt.Errorf("failed to create symlink %s", dst)
+					}
+					sylog.Debugf("Adding symlink %s to %s at %s", resolvConf, target, dst)
+
+					dest = filepath.Dir(target)
+					if err = c.session.AddDir(dest); err != nil {
+						return err
+					}
+					sylog.Debugf("Adding %s to mount list\n", dest)
+					err = system.Points.AddBind(mount.FilesTag, dest, dest, flags)
+					if err != nil {
+						return fmt.Errorf("unable to add %s to mount list: %s", dest, err)
+					}
+					c.session.OverrideDir(dest, dest)
+
+					done = true
+				}
+			}
+			if !done {
+				// regular bind mount
+				sylog.Debugf("Adding %s to mount list\n", resolvConf)
+				err = system.Points.AddBind(mount.FilesTag, resolvConf, resolvConf, flags)
+				if err != nil {
+					return fmt.Errorf("unable to add %s to mount list: %s", resolvConf, err)
+				}
+
+				err = system.Points.AddRemount(mount.FilesTag, resolvConf, flags)
 			}
 		} else {
 			dns = strings.ReplaceAll(dns, " ", "")
@@ -2266,16 +2328,16 @@ func (c *container) addResolvConfMount(system *mount.System) error {
 			if err != nil {
 				return err
 			}
-		}
-		if err := c.session.AddFile(resolvConf, content); err != nil {
-			sylog.Warningf("failed to add resolv.conf session file: %s", err)
-		}
-		sessionFile, _ := c.session.GetPath(resolvConf)
+			if err = c.session.AddFile(resolvConf, content); err != nil {
+				sylog.Warningf("failed to add resolv.conf session file: %s", err)
+			}
+			sessionFile, _ := c.session.GetPath(resolvConf)
 
-		sylog.Debugf("Adding %s to mount list\n", resolvConf)
-		err = system.Points.AddBind(mount.FilesTag, sessionFile, resolvConf, syscall.MS_BIND)
+			sylog.Debugf("Adding %s made from %s to mount list\n", resolvConf, dns)
+			err = system.Points.AddBind(mount.FilesTag, sessionFile, resolvConf, flags)
+		}
 		if err != nil {
-			return fmt.Errorf("unable to add %s to mount list: %s", resolvConf, err)
+			return fmt.Errorf("unable to add %s to mount list: %s", dest, err)
 		}
 		sylog.Verbosef("Default mount: /etc/resolv.conf:/etc/resolv.conf")
 	} else {