+++ CHANGELOG.md
@@ -5,6 +5,25 @@
 ### Bug Fixes
 
 - Don't attempt to set `relatime` on workdir / scratch mounts in OCI-Mode.
+- Include the home directory in the `--workdir` option (which is a
+  modifier of the `--contain` option).  This has always been in the
+  `--workdir` usage description but the home directory has not actually
+  been included at least since singularity-2.
+- Avoid a fatal error when starting fakeroot from suid mode while in an
+  NFS directory.
+- Support hosts that have `/etc/resolv.conf` pointing to a symlink under
+  `/run`, such as those hosts that are running `systemd-resolved`.  In
+  this case, the symlink is copied into the container and the parent
+  directory of the target of the symlink is bind-mounted from the host.
+  The result is that even if the target of the symlink is replaced with
+  a new file, the container sees the update in `/etc/resolv.conf`.
+- Correctly escape ENV vars when importing OCI containers to native SIF, so that
+  they match podman / docker behaviour.
+
+### New Features & Functionality
+
+- Add `/etc/resolv.conf` to the list of host paths that can be prevented
+  from automatic import into the container with the `--no-mount` option.
 
 ### Requirements / Packaging